fist commit

+++ src/main/java/com/mislbd/report_manager/entity/Department.java
@@ -8,23 +8,23 @@ public class Department {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "department_seq")
     @SequenceGenerator(name = "department_seq", sequenceName = "DEPARTMENT_SEQ", allocationSize = 1)
-    private Long departmentId;
+    private Integer departmentId;
 
 
     @Column(length = 100, nullable = false, unique = true)
     private String name;
 
-    public Department(Long departmentId, String name) {
+    public Department(Integer departmentId, String name) {
         this.departmentId = departmentId;
         this.name = name;
     }
 
 
-    public Long getDepartmentId() {
+    public Integer getDepartmentId() {
         return departmentId;
     }
 
-    public void setDepartmentId(Long departmentId) {
+    public void setDepartmentId(Integer departmentId) {
         this.departmentId = departmentId;
     }
